+++ lib/logux/stream.rb
@@ -12,7 +12,7 @@ module Logux
 
     def write(payload)
       processed_payload = process(payload)
-      Logux.logger_info("Write to Logux response: #{processed_payload}")
+      Logux::Logger.info("Write to Logux response: #{processed_payload}")
       stream.write(processed_payload)
     end
